move collection funcs to a separate file + minor
